[audio_to_spectogram] convert_matplotlib_to_img as a library
